Move APIs from Azure::Core::Strings to Azure::Core::Internal since they aren't meant for end users to call. (#1337)
Fixes https://github.com/Azure/azure-sdk-for-cpp/issues/1319

See https://github.com/Azure/azure-sdk-for-cpp/pull/1297#discussion_r554458422 for context.

~There is an open question whether this should be in the Jan or Feb milestone.~ Moved to Feb
