+++ sdk/core/azure-core/CMakeLists.txt
@@ -63,6 +63,7 @@ set(
     inc/azure/core/http/transport.hpp
     inc/azure/core/internal/contract.hpp
     inc/azure/core/internal/log.hpp
+    inc/azure/core/internal/strings.hpp
     inc/azure/core/logging/logging.hpp
     inc/azure/core/base64.hpp
     inc/azure/core/context.hpp
@@ -74,7 +75,6 @@ set(
     inc/azure/core/operation_status.hpp
     inc/azure/core/platform.hpp
     inc/azure/core/response.hpp
-    inc/azure/core/strings.hpp
     inc/azure/core/uuid.hpp
     inc/azure/core/version.hpp
     inc/azure/core.hpp
