+++ sdk/core/azure-core/inc/azure/core/internal/strings.hpp
@@ -2,17 +2,17 @@
 // SPDX-License-Identifier: MIT
 
 /**
- * @brief Public utility functions for strings.
+ * @brief Internal utility functions for strings.
  *
  */
 #pragma once
 
 #include <string>
 
-namespace Azure { namespace Core { namespace Strings {
+namespace Azure { namespace Core { namespace Internal { namespace Strings {
 
   bool LocaleInvariantCaseInsensitiveEqual(const std::string& lhs, const std::string& rhs) noexcept;
   std::string const ToLower(const std::string& src) noexcept;
   unsigned char ToLower(const unsigned char src) noexcept;
 
-}}} // namespace Azure::Core::Strings
+}}}} // namespace Azure::Core::Internal::Strings
