+++ sdk/core/azure-core/src/http/raw_response.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: MIT
 
 #include "azure/core/http/http.hpp"
-#include "azure/core/strings.hpp"
+#include "azure/core/internal/strings.hpp"
 
 #include <cctype>
 #include <map>
@@ -32,7 +32,7 @@ void RawResponse::AddHeader(uint8_t const* const first, uint8_t const* const las
   }
 
   // Always toLower() headers
-  auto headerName = Azure::Core::Strings::ToLower(std::string(start, end));
+  auto headerName = Azure::Core::Internal::Strings::ToLower(std::string(start, end));
   start = end + 1; // start value
   while (start < last && (*start == ' ' || *start == '\t'))
   {
