+++ sdk/core/azure-core/src/strings.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/core/strings.hpp"
+#include "azure/core/internal/strings.hpp"
 
 #include <algorithm>
 
@@ -67,7 +67,7 @@ const unsigned char c_LocaleInvariantLowercaseTable[256] = {
 };
 } // unnamed namespace
 
-namespace Azure { namespace Core { namespace Strings {
+namespace Azure { namespace Core { namespace Internal { namespace Strings {
 
   unsigned char ToLower(const unsigned char symbol) noexcept
   {
@@ -97,4 +97,4 @@ namespace Azure { namespace Core { namespace Strings {
         });
   }
 
-}}} // namespace Azure::Core::Strings
+}}}} // namespace Azure::Core::Internal::Strings
