+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -20,7 +20,6 @@ TEST(Logging, simplifiedHeader)
   EXPECT_NO_THROW(Azure::Core::Nullable<int> n);
   EXPECT_NO_THROW(Azure::Core::Http::RawResponse r(
       1, 1, Azure::Core::Http::HttpStatusCode::Accepted, "phrase"));
-  EXPECT_NO_THROW(Azure::Core::Strings::ToLower("A"));
   EXPECT_NO_THROW(Azure::Core::Uuid::CreateUuid());
   EXPECT_NO_THROW(Azure::Core::Details::Version::VersionString());
