+++ sdk/core/azure-core/test/ut/string.cpp
@@ -1,13 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <azure/core/strings.hpp>
+#include <azure/core/internal/strings.hpp>
 #include <gtest/gtest.h>
 #include <string>
 
 TEST(String, invariantCompare)
 {
-  using Azure::Core::Strings::LocaleInvariantCaseInsensitiveEqual;
+  using Azure::Core::Internal::Strings::LocaleInvariantCaseInsensitiveEqual;
   EXPECT_TRUE(LocaleInvariantCaseInsensitiveEqual("", ""));
   EXPECT_TRUE(LocaleInvariantCaseInsensitiveEqual("a", "a"));
   EXPECT_TRUE(LocaleInvariantCaseInsensitiveEqual("A", "a"));
@@ -22,7 +22,7 @@ TEST(String, invariantCompare)
 
 TEST(String, toLower)
 {
-  using Azure::Core::Strings::ToLower;
+  using Azure::Core::Internal::Strings::ToLower;
   EXPECT_TRUE(ToLower("") == "");
   EXPECT_TRUE(ToLower("a") == "a");
   EXPECT_TRUE(ToLower("A") == "a");
