+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/protocol/blob_rest_client.hpp
@@ -18,7 +18,6 @@
 #include <azure/core/http/pipeline.hpp>
 #include <azure/core/nullable.hpp>
 #include <azure/core/response.hpp>
-#include <azure/core/strings.hpp>
 #include <azure/storage/common/crypt.hpp>
 #include <azure/storage/common/storage_common.hpp>
 #include <azure/storage/common/storage_exception.hpp>
